requirements.txt
matplotlib
pytesseract
pillow
tesseract
requests
